# Add support for custom parsing of APC, SOS and PM sequences.