@kchibisov 2024-08-05
You can open a PR against alacritty with your changes, so we can use our bot for it. But generally you just build alacritty with your changes and run the `vtebench`.