@kchibisov 2024-08-05
@boazy just open a PR against alacritty's repo, it should _better_ for us to review that way, since we can run the bot.