@boazy 2024-08-07
@kchibisov I've added a pull request. I'm not sure how to trigger the benchmarks bot.