@chrisduerr 2024-08-07
> I'm not sure how to trigger the benchmarks bot.

You can't. I did.