@boazy 2024-10-16
> But couldn't you just store that on `Parser` (the size of which is irrelevant), then branch inside `Action::OpaquePut =>` and `Action::OpaqueEnd =>` since doing so should leave performance of the other branches unaffected?
> 
> I believe that's kinda your first suggestion? Why do you think this would impact performance?

I honestly don't know. I assumed the size is important for performance since we want to be able to have all state transitions in one table (which only has 256 entries?). If we go back to the start, I created this PR trying to answer the comments to the original PR that tried to add APC support (#110), particularly this comment:

>  This basically feels like an embedded parser. Most of our parser is structured in a way to stream through bytes and process them as they're coming in, however rather than creating state transitions you're storing a state transition to later then match on it instead. VTE is already too slow and I'm afraid this will just amplify that problem.

https://github.com/alacritty/vte/pull/110#discussion_r1545286222

I realize I can add additional state that doesn't affect OSC sequences, but seeing this comment I was a bit worried. Do you believe adding an additional field to `Parser` (let's say `opaque_sequence_kind`) should be ok?