@chrisduerr 2024-12-20
@boazy Just a heads-up: Once https://github.com/alacritty/vte/pull/118 is merged, it might make some sense to reinvestigate this. There's certainly been some changes and we do have a lot of extra free states/actions that could be made use of without affecting performance.