@heysweet 2025-02-03
Hello! Just wanted to re-ping this thread to see what progress was being here. We'd really like to support Kitty Image protocol in Warp, and this PR would help us get there